Move form.js and README to subdirectory
