+++ packages/web/src/index.js
@@ -3,5 +3,5 @@ export { default as RedwoodProvider, useAuth } from './RedwoodProvider'
 export { default as gql } from 'graphql-tag'
 export * from './graphql'
 export { useQuery, useMutation } from '@apollo/react-hooks'
-export * from './Form'
+export * from './form/form'
 export { default as FatalErrorBoundary } from './FatalErrorBoundary'
